Add support for the FCS flag in Radiotap.
The radio tap header can indicate whether FCS is present or not in the
recovered frames. We should expose this value to the callers. Further,
the presence of the FCS field should be propagated to the IEEE80211
module. By default, this patch sets FCS as not present, however, any
caller can indicate that FCS is icluded in the frame and it will be
parsed correctly.